Struct sqladmin1_beta4::DatabaseInstance

pub struct DatabaseInstance {
    pub max_disk_size: Option<String>,
    pub replica_names: Vec<String>,
    pub current_disk_size: Option<String>,
    pub service_account_email_address: Option<String>,
    pub ip_addresses: Vec<IpMapping>,
    pub database_version: Option<String>,
    pub instance_type: Option<String>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub ipv6_address: Option<String>,
    pub server_ca_cert: Option<SslCert>,
    pub region: Option<String>,
    pub settings: Option<Settings>,
    pub master_instance_name: Option<String>,
    pub project: Option<String>,
    pub state: Option<String>,
    pub etag: Option<String>,
    pub self_link: Option<String>,
}

A Cloud SQL instance resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch (request)

insert (request)

get (response)

update (request)

Fields

max_disk_size

The maximum disk size of the instance in bytes.

replica_names

The replicas of the instance.

current_disk_size

The current disk usage of the instance in bytes.

service_account_email_address

The service account email address assigned to the instance.

ip_addresses

The assigned IP addresses for the instance.

database_version

The database engine type and version. Can be MYSQL_5_5 or MYSQL_5_6. Defaults to MYSQL_5_5. The databaseVersion can not be changed after instance creation.

instance_type

The instance type. This can be one of the following. CLOUD_SQL_INSTANCE: A Cloud SQL instance that is not replicating from a master. READ_REPLICA_INSTANCE: A Cloud SQL instance configured as a read-replica.

kind

This is always sql#instance.

name

Name of the Cloud SQL instance. This does not include the project ID.

ipv6_address

The IPv6 address assigned to the instance.

server_ca_cert

SSL configuration.

region

The geographical region. Can be us-central, asia-east1 or europe-west1. Defaults to us-central. The region can not be changed after instance creation.

settings

The user settings.

master_instance_name

The name of the instance which will act as master in the replication setup.

project

The project ID of the project containing the Cloud SQL instance. The Google apps domain is prefixed if applicable.

state

The current serving state of the Cloud SQL instance. This can be one of the following. RUNNABLE: The instance is running, or is ready to run when accessed. SUSPENDED: The instance is not available, for example due to problems with billing. PENDING_CREATE: The instance is being created. MAINTENANCE: The instance is down for maintenance. UNKNOWN_STATE: The state of the instance is unknown.

etag

HTTP 1.1 Entity tag for the resource.

self_link

The URI of this resource.