Struct sqladmin1_beta4::ExportContext

pub struct ExportContext {
    pub csv_export_options: Option<ExportContextCsvExportOptions>,
    pub kind: Option<String>,
    pub databases: Vec<String>,
    pub file_type: Option<String>,
    pub sql_export_options: Option<ExportContextSqlExportOptions>,
    pub uri: Option<String>,
}

Database instance export context.

This type is not used in any activity, and only used as part of another schema.

Fields

csv_export_options

Options for exporting data as CSV.

kind

This is always sql#exportContext.

databases

Databases (for example, guestbook) from which the export is made. If fileType is SQL and no database is specified, all databases are exported. If fileType is CSV, you can optionally specify at most one database to export. If csvExportOptions.selectQuery also specifies the database, this field will be ignored.

file_type

The file type for the specified uri. SQL: The file contains SQL statements. CSV: The file contains CSV data.

sql_export_options

Options for exporting data as SQL statements.

uri

The path to the file in Google Cloud Storage where the export will be stored. The URI is in the form gs://bucketName/fileName. If the file already exists, the operation fails. If fileType is SQL and the filename ends with .gz, the contents are compressed.